refactor(widget): 重构标签打印功能
- 优化了标签打印的逻辑和布局
- 新增了固定单码标和合并动态标签的打印功能
- 改进了表格的拆分和显示方式
- 调整了不同语言的显示格式

+++ lib/bean/http/response/sap_wms_reprint_label_info.dart
@@ -14,7 +14,7 @@ class ReprintLabelInfo {
   String? instructionNo; //指令号 ZVBELN_ORI
   String? salesOrderNumber; //销售订单号  KDAUF
   int? salesOrderLineItem; //销售订单行号  KDPOS
-  double? quantity; //数量 MENGE
+  int? quantity; //数量 MENGE
   String? unit; // MEINS
   String? deliveryDate; //单位 EINDT
   String? numPage; //交货日期  ZPQYM
